requirements.txt
python-telegram-bot==21.6
scikit-learn==1.5.2
numpy==1.26.4
pymorphy3==2.0.2
sentence-transformers==3.3.1
matplotlib==3.9.2
